+++ panel-no-op/src/main/kotlin/com/redmadrobot/debug/noop/plugin/aboutapp/AboutAppInfo.kt
@@ -1,9 +1,13 @@
-package com.redmadrobot.debug.plugin.aboutapp
+package com.redmadrobot.debug.plugin.aboutapp.model
+
+import java.util.UUID
 
 /**
  * No-op declaration of [AboutAppInfo] for release builds.
  */
 public data class AboutAppInfo(
     val title: String,
-    val value: String
-)
+    val value: String,
+) {
+    val id: String = UUID.randomUUID().toString()
+}