+++ sample/src/release/kotlin/com/redmadrobot/debug_sample/debug_data/DebugAboutAppInfoProvider.kt
@@ -1,6 +1,6 @@
 package com.redmadrobot.debug_sample.debug_data
 
-import com.redmadrobot.debug.plugin.aboutapp.AboutAppInfo
+import com.redmadrobot.debug.plugin.aboutapp.model.AboutAppInfo
 
 class DebugAboutAppInfoProvider {
     companion object {